backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.2.0
stellar-sdk==13.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
gunicorn==21.2.0
PyJWT==2.8.0